requirements.txt
alembic==1.4.2
appdirs==1.4.4
Babel==2.8.0
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorclass==2.2.0
distlib==0.3.1
docopt==0.6.2
dominate==2.5.1
elasticsearch==7.8.0
filelock==3.0.12
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
guess-language-spirit==0.5.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.4
pbr==5.4.5
pip-upgrader==1.4.15
psycopg2
pur==5.3.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
stevedore==3.1.0
terminaltables==3.1.0
urllib3==1.25.9
virtualenv==20.0.26
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
